- enable ctrl-c to cancel the test safely. - add tests. (gfarmfs_attr0.out, gfarmfs_oldio.out, gfarmfs_oldio_F.out) - save errlog of gfarmfs. (not yet compare errlog) - use GFARM_PATH_INFO_TIMEOUT=1000 (allways effective) - gfarmfs -nlsu is default. - Test of oldio use attr_timeout=0.

+++ gfarmfs-fuse/trunk/test/test.sh
@@ -9,6 +9,8 @@ EXPECTDIR=./expected
 FSYSTEST=./fsystest
 GFARMFS=../gfarmfs
 
+export GFARM_PATH_INFO_TIMEOUT=1000
+
 ##### test mode #####
 TESTMODE=$1
 DO_TMP=0
@@ -19,6 +21,10 @@ if [ x"${TESTMODE}" = x"all" ]; then
     DO_TMP=1
     DO_GFARMFS=1
     DO_FUSEXMP=1
+elif [ x"${TESTMODE}" = x"tmp" ]; then
+    DO_TMP=1
+    DO_GFARMFS=0
+    DO_FUSEXMP=0
 elif [ x"${TESTMODE}" = x"gfarmfs" ]; then
     DO_TMP=0
     DO_GFARMFS=1
@@ -28,12 +34,19 @@ elif [ x"${TESTMODE}" = x"fusexmp" ]; then
     DO_GFARMFS=0
     DO_FUSEXMP=1
 else
-    echo "usage: $0 <all|gfarmfs|fusexmp>"
+    echo "usage: $0 <all|gfarmfs|fusexmp|tmp>"
     exit 1
 fi
 
 ##### functions #####
 err=0
+canceled=0
+
+test_cancel()
+{
+    echo "canceled (test.sh)"
+    canceled=1
+}
 
 init()
 {
@@ -52,6 +65,8 @@ init()
         exit 1
     fi
     err=0
+    trap test_cancel 1
+    trap test_cancel 2
 }
 
 check_result()
@@ -61,10 +76,10 @@ check_result()
     IGNORE=$3
 
     if [ $RESULT -ne 0 -a $IGNORE -eq 1 ]; then
-        echo "IGNORE(${RESULT}): ${TESTSTR}" 1>&2
+        echo "IGNORE(${RESULT}): ${TESTSTR}"
     elif [ $RESULT -ne 0 ]; then
         err=`expr $err + 1`
-        echo "NG(${RESULT}): ${TESTSTR}" 1>&2
+        echo "NG(${RESULT}): ${TESTSTR}"
     else
         echo "OK(${RESULT}): ${TESTSTR}"
     fi
@@ -73,13 +88,22 @@ check_result()
 test_common()
 {
     TESTDIR=$1
-    OUTPUT=${OUTPUTDIR}/${2}
-    DIFFOUT=${DIFFDIR}/${2}
-    EXPECTED=$3
+    OUTPUT=${OUTPUTDIR}/${2}.out
+    DIFFOUT=${DIFFDIR}/${2}.out
+    EXPECTEDOUT=${EXPECTDIR}/${3}.out
     TESTSTR=$4
 
-    ${FSYSTEST} ${TESTDIR} > ${OUTPUT} 2>&1
-    diff -u ${EXPECTDIR}/${EXPECTED} ${OUTPUT} > ${DIFFOUT} 2>&1
+    if [ $canceled -eq 1 ];then
+        return
+    fi
+    ${FSYSTEST} ${TESTDIR} > ${OUTPUT} 2>&1 &
+    testpid=$!
+    wait $testpid
+    if [ $canceled -eq 1 ];then
+        wait $testpid
+#        cat ${OUTPUT}
+    fi
+    diff -u ${EXPECTEDOUT} ${OUTPUT} > ${DIFFOUT} 2>&1
     result=$?
     if [ $result -eq 0 ]; then
         rm ${DIFFOUT} > /dev/null 2>&1
@@ -114,8 +138,7 @@ fuse_common_init()
     TESTNAME=$1
     OUTNAME=$2
 
-    EXPECTED=$OUTNAME
-    OUTPUT=${OUTPUTDIR}/${OUTNAME}
+    OUTPUT=${OUTPUTDIR}/${OUTNAME}.out
     umount_fuse ${TMP_MNTDIR} > /dev/null 2>&1
     rmdir ${TMP_MNTDIR} > /dev/null 2>&1
     if [ -e ${TMP_MNTDIR} ]; then
@@ -132,7 +155,8 @@ fuse_common_do_test()
     TESTDIR=$2
     OUTNAME=$3
 
-    OUTPUT=${OUTPUTDIR}/${OUTNAME}
+    EXPECTED=${OUTNAME}
+    OUTPUT=${OUTPUTDIR}/${OUTNAME}.out
     if [ $RESULT -ne 0 ];then
         check_result -1 "${TESTNAME}: `cat ${OUTPUT}`" 0
         return $RESULT
@@ -159,17 +183,20 @@ test_gfarmfs()
     FUSEOPTIONS=$2
     OUTNAME=$3
 
-    EXPECTED=$OUTNAME
-    OUTPUT=${OUTPUTDIR}/${OUTNAME}
-    TESTNAME="gfarmfs: opt=\"$1\", fuse_opt=\"$2\""
+    OUTPUT=${OUTPUTDIR}/${OUTNAME}.out
+    ERRLOGOPT="--errlog ${OUTPUTDIR}/errlog-${OUTNAME}"
+    TESTNAME="gfarmfs: opt=\"${OPTIONS}\", fuse_opt=\"${FUSEOPTIONS}\""
     TESTDIR=${TMP_MNTDIR}/${LOGNAME}  # need gfmkdir gfarm:~
 
+    if [ $canceled -eq 1 ];then
+        return
+    fi
     fuse_common_init "${TESTNAME}" ${OUTNAME}
     result=$?
     if [ $result -ne 0 ];then
         return $result
     fi
-    $GFARMFS ${OPTIONS} ${TMP_MNTDIR} ${FUSEOPTIONS} > ${OUTPUT} 2>&1
+    $GFARMFS $OPTIONS $ERRLOGOPT $TMP_MNTDIR $FUSEOPTIONS > $OUTPUT 2>&1
     result=$?
     fuse_common_do_test $result ${TESTDIR} ${OUTNAME}
     fuse_common_final
@@ -181,11 +208,13 @@ test_fusexmp()
     FUSEOPTIONS=$2
     OUTNAME=$3
 
-    EXPECTED=$OUTNAME
-    OUTPUT=${OUTPUTDIR}/${OUTNAME}
+    OUTPUT=${OUTPUTDIR}/${OUTNAME}.out
     TESTNAME="$FUSEXMP: fuse_opt=\"${FUSEOPTIONS}\""
     TESTDIR=${TMP_MNTDIR}/tmp
 
+    if [ $canceled -eq 1 ];then
+        return
+    fi
     which $FUSEXMP 2> $OUTPUT 1> /dev/null
     result=$?
     if [ $result -ne 0 ]; then
@@ -208,27 +237,29 @@ init
 
 ##### test on /tmp #####
 if [ $DO_TMP -eq 1 ]; then
-    test_common /tmp slashtmp.out slashtmp.out "/tmp" 
+    test_common /tmp slashtmp slashtmp "/tmp" 
 fi
-
 
 ##### test on gfarmfs #####
 if [ $DO_GFARMFS -eq 1 ]; then
-    test_gfarmfs "" "" gfarmfs_noopt.out
-    test_gfarmfs "-nlsu" "" gfarmfs_nlsu.out
-    test_gfarmfs "-N2" "" gfarmfs_N2.out
-    test_gfarmfs "-b" "" gfarmfs_b.out
-    test_gfarmfs "-b" "-o direct_io" gfarmfs_b_direct_io.out
-    test_gfarmfs "" "-o direct_io" gfarmfs_direct_io.out
+    test_gfarmfs "" "" gfarmfs_noopt
+    test_gfarmfs "-nlsu" "" gfarmfs_nlsu
+    test_gfarmfs "-nlsu" "-o attr_timeout=0" gfarmfs_attr0
+    test_gfarmfs "-nlsu -N2" "" gfarmfs_N2
+    test_gfarmfs "-nlsu -b" "" gfarmfs_b
+    test_gfarmfs "-nlsu -b" "-o direct_io" gfarmfs_b_direct_io
+    test_gfarmfs "-nlsu" "-o direct_io" gfarmfs_direct_io
+    test_gfarmfs "--oldio -nlsub" "-o attr_timeout=0" gfarmfs_oldio
+    test_gfarmfs "--oldio -nlsubF" "-o attr_timeout=0" gfarmfs_oldio_F
 fi
 
 ##### fusexmp_fh #####
 if [ $DO_FUSEXMP -eq 1 ]; then
-    test_fusexmp "fusexmp" "" fusexmp.out
-    test_fusexmp "fusexmp_fh" "" fusexmp_fh.out
-    test_fusexmp "fusexmp_fh" "-o direct_io" fusexmp_fh_direct_io.out
-    test_fusexmp "fusexmp_fh" "-o kernel_cache" fusexmp_fh_kernel_cache.out
-    test_fusexmp "fusexmp_fh" "-s" fusexmp_fh_s.out
+    test_fusexmp "fusexmp" "" fusexmp
+    test_fusexmp "fusexmp_fh" "" fusexmp_fh
+    test_fusexmp "fusexmp_fh" "-o direct_io" fusexmp_fh_direct_io
+    test_fusexmp "fusexmp_fh" "-o kernel_cache" fusexmp_fh_kernel_cache
+    test_fusexmp "fusexmp_fh" "-s" fusexmp_fh_s
 fi
 
 ##### final #####
